refactor: use priorities constant in task creation form select element
