+++ src/app/tasks/components/columns.tsx
@@ -2,12 +2,12 @@
 
 import { ColumnDef } from '@tanstack/react-table';
 import { Checkbox } from '@/components/ui/checkbox';
-import { PRIORITIES } from '../constants';
+import { PRIORITIES } from '@/constants';
 import { DataTableColumnWithSorting } from './data-table-column-with-sorting';
 import { DataTableRowActions } from './data-table-row-actions';
 import { DataTableCellWithIcon } from './data-table-cell-with-icon';
 import { FunctionComponent } from 'react';
-import { Task } from '../types';
+import { Task } from '@/types';
 
 export const columns: ColumnDef<Task>[] = [
     {
